# Serverless-Finance/lib/api-zod/src/generated/types/portfolioSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaVest API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioSummary {
  totalWealth: number;
  activePrincipal: number;
  totalROIReceived: number;
  liquidity: number;
  activeInvestmentCount: number;
}