# Serverless-Finance/lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaVest API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  fullName?: string;
  theme?: string;
  biometricEnabled?: boolean;
  bankName?: string;
  bankAccountNumber?: string;
  bankAccountName?: string;
  cryptoWithdrawAddress?: string;
  cryptoWithdrawNetwork?: string;
}