# Serverless-Finance/lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AlphaVest API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TransactionInput {
  type: string;
  fund: string;
  amount: number;
}